requirements.txt
accelerate==0.25.0
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cmake==3.28.1
decord==0.6.0
diffusers==0.25.0
docopt==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.12.2
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.40
hope
huggingface-hub==0.20.1
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
inquirerpy==0.3.4
Jinja2==3.1.2
jsonlines==4.0.0
lightning-utilities==0.10.0
lit==17.0.6
MarkupSafe==2.1.3
mpi4py
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
pfzy==0.3.4
Pillow==10.1.0
pluggy==1.3.0
ply==3.11
prettytable==3.9.0
prompt-toolkit==3.0.41
protobuf==3.20.3
psutil==5.9.7
pyre-extensions==0.0.29
pytest==7.4.3
python-dateutil==2.8.2
pytorch-lightning==2.1.3
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
six==1.16.0
smmap==5.0.1
sympy==1.12
tfrecord==1.14.4
thrift==0.16.0
thriftpy2==0.4.16
tokenizers==0.15.0
tomli==2.0.1
torch==2.0.0
torchmetrics==1.2.1
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
xformers==0.0.19
yarl==1.9.4
zipp==3.17.0
lpips
